fix: fix up publishing to Kafka
There were a few issues with the Kafka producer - firstly, to enable it
required setting 'KAFKA_ENABLED=false' due to an incorrect comparison
with 'false'. The kafkajs library requires setting the partitioner value
to not throw an error.

I found it was necessary to modify some of the connection logic and add
some async/await defs because we were publishing to a connection that
was being closed, resulting in an error and subsequent failure to
publish messages.